backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery[redis]==5.4.0
redis==5.0.8
django-cors-headers==4.4.0
whitenoise==6.7.0
gunicorn==22.0.0
python-dotenv==1.0.1
